refactor code, use seperate controller file for each controller feature
